requirements_dev.txt
# Generic python dependencies
black==25.11.0
mypy==1.18.2
pip==25.3
pre-commit==4.5.0
ruff==0.14.6

# Home Assistant testing dependencies
pytest-homeassistant-custom-component==0.13.297
syrupy>=4.6.1

# Component dependencies
hassil>=3.2.0
home-assistant-intents>=2025.9.24
supernote==0.3.1
